ai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
web3==7.5.0
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.2
pydantic==2.10.0
